git-svn-id: file:///srv/local/data/cleanRepo/trunk@16167 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/wrashid/src/main/java/playground/wrashid/parkingChoice/infrastructure/api/Parking.java
@@ -11,9 +11,9 @@ public interface Parking  extends Identifiable {
 	
 	public abstract Coord getCoord();
 	
-	public abstract int getCapacity();
+	public abstract double getCapacity();
 	
-	public abstract void setCapacity(int capacity);
+	public abstract void setCapacity(double capacity);
 	
 	public abstract PriceScheme getPriceScheme();
 }
